--- backend/evaluation/eval_auto_results_improved.csv
@@ -1,37 +1,0 @@
-article_id,mode,redundancy,keyword_coverage,rouge1_f,rouge2_f,rougeL_f
-1,textrank,0.0,0.4,0.0,0.0,0.0
-1,kobart,0.0,0.2,0.4444444444444445,0.2857142857142857,0.4444444444444445
-1,hybrid,0.0,0.6,1.0,1.0,1.0
-2,textrank,0.0,0.6,0.0,0.0,0
-2,kobart,0.0,0.6,0.0,0.0,0
-2,hybrid,0.0,0.6,0.0,0.0,0
-3,textrank,0.0,0.4,0.3684210526315789,0.16666666666666669,0.3684210526315789
-3,kobart,0.0,0.4,0.25,0.10526315789473684,0.25
-3,hybrid,0.0,0.4,0.3225806451612903,0.13793103448275862,0.3225806451612903
-4,textrank,0.0,0.4,0.23076923076923078,0.17647058823529413,0.23076923076923078
-4,kobart,0.0,0.0,0.45000000000000007,0.3157894736842105,0.45000000000000007
-4,hybrid,0.0,0.0,0.45000000000000007,0.3157894736842105,0.45000000000000007
-5,textrank,0.0,0.0,0.23076923076923078,0.16666666666666669,0.23076923076923078
-5,kobart,0.0,0.0,0.3333333333333333,0.2,0.3333333333333333
-5,hybrid,0.0,0.0,0.3333333333333333,0.2,0.3333333333333333
-6,textrank,0.0,0.6,0.0,0.0,0
-6,kobart,0.0,0.2,0.0,0.0,0
-6,hybrid,0.0,0.2,0.0,0.0,0
-7,textrank,0.0,0.0,0.8421052631578948,0.588235294117647,0.8421052631578948
-7,kobart,0.0,0.0,0.5925925925925926,0.24,0.5185185185185185
-7,hybrid,0.0,0.0,0.5925925925925926,0.24,0.5185185185185185
-8,textrank,0.0,0.8,0.19999999999999998,0.0,0.19999999999999998
-8,kobart,0.0,0.8,0.2857142857142857,0.0,0.2857142857142857
-8,hybrid,0.0,0.8,0.6666666666666666,0.0,0.6666666666666666
-9,textrank,0.0,1.0,0.3870967741935483,0.20689655172413793,0.3870967741935483
-9,kobart,0.0,0.6,0.5714285714285714,0.3157894736842105,0.5714285714285714
-9,hybrid,0.0,0.8,0.5263157894736842,0.3529411764705882,0.5263157894736842
-10,textrank,0.0,0.8,0.4324324324324324,0.2857142857142857,0.32432432432432434
-10,kobart,0.0,0.6,0.4615384615384615,0.18181818181818182,0.3076923076923077
-10,hybrid,0.0,0.6,0.4615384615384615,0.18181818181818182,0.3076923076923077
-11,textrank,0.0,0.8,0.2222222222222222,0.125,0.2222222222222222
-11,kobart,0.0,0.4,1.0,1.0,1.0
-11,hybrid,0.0,0.4,1.0,1.0,1.0
-12,textrank,0.0,0.4,0.5714285714285715,0.4,0.5714285714285715
-12,kobart,0.0,0.4,0.0,0.0,0.0
-12,hybrid,0.0,0.4,0.0,0.0,0.0
